Refresh region feature splatting values container
We have made major changes to the implementation of the container for storing region feature splatting values, namely the *splatting value matrix*. The API has changed significantly with deep integration with the STL range library, rather than solely relying on multi-dimensional random-accessed indexing. As of current, the range accessor flattens the two spatial dimensions and only keeps the per-region splatting values; we may add a *md-range* in the future if necessary.

Implementation of comparison operators are greatly simplified, as we have previously moved the inter-conversion between dense and sparse matrix to `SparseMatrixElement` as range adaptors.

For consistency, importance computer has also been renamed to splatting.

> P.S. Forgot to mention in the [last commit](https://github.com/stephen-hqxu/drr/commit/bb716bcc200bb61d5335a0fdf38b1aed67332ed6), `Combinator` has been renamed to `Splatting`. This is done because we found it is more analogous to another well-established concept called *texture splatting*.