Split project into three build targets
The original `Main` target is split into:
- Library
- Programme
- Test
We believe it is more logically sensible to move testing code to a separate executable, potentially with use of a unit-test framework. In that sense, the core part of the project would need to be built into a static library. This also allows package dependencies refactoring as the main library does not need any of the CLI or benchmarking library.

Thread pool and system control are moved from the old *Profiler* to *Core*. The remains goes to *Programme* and *Test* wherever appropriate in project root.

To allow git tracking histories of files, this commit only renames them and does not update the build system. It is expected that CMake would not run at all.