Build script refactor
Functions for setting up target sources and source groups are moved to project root, and they are made to be more generalised so can be used for setting up other targets in this project.

Instead of using hard-coded target names, as it is inconvenient to change wherever it is referenced when it is renamed, we adopt a similar approach as CMake's built-in variables like current source directory for currently processed target and its include directory, since we only ever deal with one target per subdirectory added to the project root.

For customised external targets (like *xxHash*), add an alias target with project name as prefix to disambiguate.