All splatting implementations are now stateless
The main reason is to facilitate the implementation of parallel splatting profiler. Using a stateful splatting, e.g. radius is stored as a member, requires special precautions to prevent data racing, such as the user needs to create separate objects for each executing thread.

Although this forces us to make `InvokeInfo` polymorphic and use the ugly `dynamic_cast`. It should not affect the performance at all, if there is any, as all ugly stuff are outside the high traffic zone, and casting are only done a few times before the actually run.

Regionfield input is removed from `InvokeInfo` and passed directly to the argument for stylistic purposes, because it is more related to the splatting operation.