Make convolution-based splatting stateful again
We found out when writing splatting profiler that, although having a stateless splatting class allows profiling the same splatting instance with different input configurations on different threads, such as when varying radius of a convolution-based splatting, there is no intuitive way of setting fixed parameters, e.g. when profiling something else and we need a fixed radius, as the profiler has no knowledge on the concrete implementation of a splatting. There is also no easy way of allowing the user to set the fixed parameters because that would involve doing some type-erasure magic to pass such parameters to a derived `InvokeInfo`.

It is probably better to stick to the separate objects per-thread idea, which should not be too bad because all splatting class are fairly lightweight and trivial (in terms of memory layout).