Change to storage order and coordinate system convention
To make iterating through the matrix more intuitive, we changed the mdspan layout for regionfield and splatting coefficient to `std::layout_right` from `std::layout_left`. This is to eliminate the possibility of human mistakes since, when indexing into the matrix, the X axis is first specified followed by the Y axis. This creates a frustration when forming a nested loop and can easily loop in the wrong order that ruins cache locality.

Therefore, the Y axis now gets a stride of one, and one can naturally form a nested loop by having X being the outer loop, and so on.

For the splatting coefficient matrix, we have also move the region count axis to the right-most axis, so it has a stride of one.

# New core view for refactor
- MdSpan: Add to refactor commonly used utilities on `std::mdspan`, including conversion between `glm::vec`.
- Functional.MakeFromTuple: A wrapper for `std::views::transform` with a `std::make_from_tuple` in its body.

# Others
- Replace all uses of `std::index_sequence` with `std::integer_sequence` with explicit type that matches the index type of the container being accessed.
- ThreadPool: Fix a compilation error for ambiguous implicit overload resolution to priority and affinity mask setting functions.
- Test.StringMaker: Remove specialisation to `std::formatter` and implement `Catch::StringMaker` directly. Add one for `SparseMatrixElement::Basic` to assist debugging.