High-level regionfield generator entry point
Programme.Generator.Regionfield is a high-level entry point to generate a regionfield matrix using one of the built-in regionfield generator implemented in the main library. This is to prepare for the later integration of command line regionfield generator tool.

For readability, regionfield generator execution policy traits are all initialised into a global variable, instead of keep doing the inline brace initialisation with the long `DRR_REGIONFIELD_GENERATOR_EXECUTION_POLICY_TRAIT` macro.