Fix container serialisation layout
An error is generated when writing a non-square regionfield and region splatting coefficient matrix. This is because of the output is written transposed but the image dimension is not.

Also fixed an error when writing a SCM to TIFF for using an incorrect tile dimensionality. The tile is now formatted as a 2D matrix, and masks are written region by region.

The `flip` function in Core.MdSpan has been renamed to `reverse`. For convenience of implementation, the function now takes a GLM vector instead of a std extents.