High-level region feature spaltting coefficient computing
A high-level function is provided to compute region feature splatting coefficients, given a regionfield matrix, using a range of implemented splatting algorithms, and obtain a dense region mask.

For the high-level regionfield generator, the seed in `GenerateInfo` is replaced with the generate info from the base regionfield generator for abstraction.

# Splatting
- Base: Add a new query function in the base splatting for getting the maximum splatting extent, given a known regionfield dimension and splatting offset. Add additional input validations.
- Container: Initialise all valid combinations of splatting container trait as global constants, similar to what has been done to regionfield generator execution policy traits in commit 813285657c15bb4a8abdd9af9bfcdeed71e401e6. A tuple of all trait combinations are also provided, as both a type alias and a global constant.