Make stochastic splatting executes faster
The previous implementation of generating a 2D index sequence and shuffling runs extremely slow, as its runtime depends on the kernel size. For sampled convolutions, its runtime should be independent to the kernel size. The new implementation simply generates random sample indices without worrying about taking duplicate samples.

The CLI is improved for maintainability, and allow specifying compression algorithms with flags rather than enum. This is mainly because the compression level configuration depends on the algorithm used, such as its default value which is 3 for zstd but may be different for others.