New regionfield generator with Diamond-Square algorithm
This is an algorithm originally used for procedural terrain generation. We have re-engineered it for regionfield. The algorithm is redesigned with decent instruction-and-data-level parallelism.

New features are added to our XXHash implementation. It now allows application to create a secret from string of arbitrary length. When hashing an object, the object size and address are obtained strategically instead of always using `sizeof` and `std::addressof`. For example, if the object is a `std::array`, it is more desirable to query the array size and underlying data pointer.