Command line utility for plotting splatting profiler results
It has come to the awareness that the old *Documentation/drr.ipynb* is quite inconvenient to use, as its main functionality is for the visualisation and serialisation of profiler results. This is now been migrated to Python scripts with properly designed command line interface, similar to that of the *drr* programme.