Allow specifying masked feature data type and rescaling
Output data type override is mainly designed for saving to an image format differ from the input. For example, integer underflow may occur if the destination format does not support signed integer.

Rescaling is mainly to facilitate visualisation if the features are small in variation (e.g. DEM). It will modify the raw data thus should not be used in applications that rely on data originality.