Split dense region feature splatting mask
The dense mask output by the *drr* programme is a multi-directory volumetric TIFF image. This proves significant difficulties to a majority of image reader and often requires custom tools to open this image (e.g. *tifffile*). This commit introduces a new plotting engine for splitting a complex dense mask into a collection of simple images, where each consists of a single directory and depth.

- Refactor image format command line arguments to a new submodule, `ArgumentGroup`.
- For `masked` and `split`, output image extension CLI has changed. Should now use *--format* instead of *--extension*. The image file extension is deduced automatically based on the format.
- Explicitly specify argument type for most arguments, especially for `Path`.